requirements.txt
selenium==3.141.0
aiogram
telegram==0.0.1
decorator==4.4.2
future==0.18.2
python-telegram-bot==12.6.1
tornado==6.0.4
requests==2.7.0
fake-useragent==0.1.11
beautifulsoup4==4.8.2
psycopg2-binary==2.8.4